gallery.js last part
after the array has been edited i had to update the element with document.getElementById(" ").innerHTML. then make the innerHTML = to the random array by using (current.name of array). only one differnece is for the picture. for picture  i had to use (src) instead of (innerHTML). and for source i had type it in ad  ("image courtesy of:" + current.form) so it will write it as  image courtesy of: restaurantschico.com or whatever the array of from is.